fix bug in reconstruction & get right id codes for INDRA
acq params were reset in KVEventReconstructor::ReconstructEvent by call to
KVDetectorEvent::Clear before new event was reconstructed.
this meant if the same group was in the previous and current event,
its acq params were set to -1, leading to the bizarre case where a group
appeared in the list of hit groups but with no fired acq params.

add KVIDTelescope argument to KVReconstructedNucleus::SetIdentification,
this allows call to KVIDTelescope::SetIDCode which has already been
cunningly overridden in KVINDRAIDTelescope